+++ Model/Helpers.swift
@@ -1,0 +1,16 @@
+//
+//  Helpers.swift
+//  Cuisine101
+//
+//  Created by Rayane Xavier on 02/04/20.
+//  Copyright © 2020 Nathalia Melare. All rights reserved.
+//
+
+import UIKit
+
+enum CategoryOfRenevue: String {
+    case cooked = "Cooked"
+    case frying = "Frying"
+    case roast = "Roast"
+    case drink = "Drink"
+}
